Fixed butterfly bush
I've used a function to convert strings to camel case, to fix the bug I had where flower with spaces in them wouldn't show up. It would also be good to test if flowers with hyphens work.

I can now start adding more months with the tasks in them for each flower.